+++ lib/features/password_manager/store_settings/widgets/pinned_entity_types_modal.dart
@@ -7,7 +7,7 @@ import 'package:hoplixi/core/utils/toastification.dart';
 import 'package:hoplixi/features/password_manager/dashboard/dashboard.dart';
 import 'package:hoplixi/features/password_manager/store_settings/providers/pinned_entity_types_provider.dart';
 import 'package:hoplixi/features/password_manager/store_settings/widgets/pinned_entity_types_selector.dart';
-import 'package:hoplixi/main_db/config/store_settings_keys.dart';
+import 'package:hoplixi/main_db/core/config/store_settings_keys.dart';
 import 'package:hoplixi/main_db/providers/other/dao_providers.dart';
 import 'package:hoplixi/shared/ui/button.dart';
 import 'package:hoplixi/shared/ui/modal_sheet_close_button.dart';