+++ lib/main_db/core/services/history/store_history_policy_service.dart
@@ -1,4 +1,4 @@
-import 'package:hoplixi/main_db/config/store_settings_keys.dart';
+import 'package:hoplixi/main_db/core/config/store_settings_keys.dart';
 import 'package:hoplixi/main_db/core/daos/base/system/store_settings_dao.dart';
 
 class StoreHistoryPolicyService {