+++ lib/main_db/usecases/perform_store_cleanup.dart
@@ -1,6 +1,6 @@
 import 'package:hoplixi/core/logger/app_logger.dart';
 import 'package:hoplixi/main_db/core/old/daos/daos.dart';
-import 'package:hoplixi/main_db/config/store_settings_keys.dart';
+import 'package:hoplixi/main_db/core/config/store_settings_keys.dart';
 import 'package:hoplixi/main_db/services/other/file_storage_service.dart';
 
 enum StoreCleanupStatus { completed, skippedByInterval, failed }